+++ src/index.ts
@@ -135,6 +135,10 @@ import IFreeable from "./pooling/ifreeable";
 import ArraySystem from "./system/array_system";
 import StatefulSystem from "./system/stateful_system";
 import IJamJarGlobals from "./jamjar_globals";
+import CanvasResizeSystem from "./standard/canvas_resize/canvas_resize_system";
+import CanvasResize from "./standard/canvas_resize/canvas_resize";
+import FakeResizeObserver from "./fake/fake_resize_observer";
+import FakeScreen from "./fake/fake_screen";
 
 export {
     // Core
@@ -205,6 +209,8 @@ export {
     FakeScene,
     FakeSubscriber,
     FakeWebGL2RenderingContext,
+    FakeResizeObserver,
+    FakeScreen,
     // Standard Lib
     // Audio Source
     AudioSource,
@@ -284,6 +290,9 @@ export {
     DefaultTextureFragmentShader,
     DefaultTextureVertexShader,
     WebGLSystem,
+    // Canvas resizing
+    CanvasResizeSystem,
+    CanvasResize,
 };
 
 declare global {